feat(comment-timezone): Format timestamp and convert to local in browser
The timestamp saved in the database is UTC based on my testing and what
I could find online. The SQLite documentation was a bit vauge around this.

Since the timestamp being returned is already UTC I am just converting in the
browser, since it will use the client's local timezone.

I'va also adjusted things to format the timestamp in the browser instead of on
the server. This is allowing us to take advantage of that conversion to local time